Avatar updated once database object changes
