+++ Zom/Zom/Classes/View Controllers/ZomProfileViewController.swift
@@ -9,6 +9,7 @@
 import Foundation
 import PureLayout
 import OTRAssets
+import MobileCoreServices
 
 extension OTRBuddy {
     func zom_inviteLink(otrFingerprint:String?) -> NSURL {
@@ -22,55 +23,9 @@ extension OTRBuddy {
     }
 }
 
-/** All the idenitifiers for each cell that is possible in a ProfileViewController */
-internal enum ZomProfileViewCellIdentifier:String {
-    case ProfileCell = "ProfileCell"
-    case FingerprintCell = "FingerprintCell"
-    case ButtonCell = "ButtonCell"
-    case PasswordCell = "PasswordCell"
-    
-    static let allValues = [ProfileCell,FingerprintCell,ButtonCell,PasswordCell]
-    
-    enum ClassOrNib {
-        case Nib(UINib)
-        case Class(AnyClass)
-    }
-    
-    /** Get the cell class or nib depending on the identifier type */
-    static func classOrNib(identifier:ZomProfileViewCellIdentifier) -> ClassOrNib {
-        let resourceBundle = OTRAssets.resourcesBundle()
-        switch identifier {
-        case .ProfileCell :
-            return .Nib(UINib(nibName: "ZomUserInfoProfileCell", bundle: resourceBundle))
-        case FingerprintCell:
-            return .Nib(UINib(nibName: "ZomFingerprintCell", bundle: resourceBundle))
-        case PasswordCell :
-            return .Nib(UINib(nibName: "ZomPasswordCell", bundle: resourceBundle))
-        case .ButtonCell:
-            return .Class(UITableViewCell.self)
-        }
-    }
-}
 
-/** 
- This protocol defines the funtion for any cell info struct/object.
- The Cell info should contain all the data necesssary to build the UITableViewCell of its type.
- */
-protocol ZomProfileViewCellInfoProtocol {
-    
-    /** 
-     Called from `func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell`.
-     */
-    func configure(cell:UITableViewCell)
-    
-    /** The cell type for this cell info */
-    func cellIdentifier() -> ZomProfileViewCellIdentifier
-    
-    /** The cell height. If nil then UITableViewAutomaticDimension is used */
-    func cellHeight() -> CGFloat?
-}
 
-/** This contains all teh information necessary to build the ZomProfileViewController */
+/** This contains all the information necessary to build the ZomProfileViewController */
 struct ZomProfileViewControllerInfo {
     
     struct zomOTRKitInfo {
@@ -98,6 +53,20 @@ struct ZomProfileViewControllerInfo {
                 })
                 return
             }
+        }
+        
+        func databaseObject() -> OTRYapDatabaseObject {
+            switch self {
+            case Buddy(let buddy): return buddy
+            case Account(let account): return account
+            }
+        }
+        
+        func yapKey() -> String {
+            return self.databaseObject().uniqueId
+        }
+        func yapCollection() -> String {
+            return self.databaseObject().dynamicType.collection()
         }
     }
     
@@ -179,139 +148,9 @@ struct ZomProfileViewControllerInfo {
     }
 }
 
-/** This struct contains all the information for a table section  */
-struct TableSectionInfo {
-    /** The title of the section */
-    let title:String?
-    /** The cells in the section */
-    let cells:[ZomProfileViewCellInfoProtocol]?
-}
-
-/** Contains all the information necessary to render the user cell */
-struct UserCellInfo: ZomProfileViewCellInfoProtocol {
-    
-    let avatarImage:UIImage?
-    let title:String
-    let subtitle:String?
-    
-    static let kCellHeight:CGFloat = 90
-    
-    func configure(cell: UITableViewCell) {
-        guard let userCell = cell as? ZomUserInfoProfileCell else {
-            return
-        }
-        
-        userCell.displayNameLabel.text = self.title
-        userCell.usernameLabel.text = self.subtitle
-        userCell.avatarImageView.setImage(self.avatarImage, forState: .Normal)
-        userCell.avatarImageView.layer.cornerRadius = CGRectGetWidth(userCell.avatarImageView.frame)/2;
-        userCell.avatarImageView.userInteractionEnabled = true
-        userCell.avatarImageView.clipsToBounds = true;
-        userCell.selectionStyle = .None
-    }
-    
-    func cellIdentifier() -> ZomProfileViewCellIdentifier {
-        return .ProfileCell
-    }
-    
-    func cellHeight() -> CGFloat? {
-        return UserCellInfo.kCellHeight
-    }
-}
-
-struct FingerprintCellInfo: ZomProfileViewCellInfoProtocol {
-    
-    let fingerprint:String
-    let qrAction:((info:FingerprintCellInfo)->Void)?
-    let shareAction:((info:FingerprintCellInfo)->Void)?
-    private let shareImage = UIImage(named: "OTRShareIcon", inBundle: OTRAssets.resourcesBundle(), compatibleWithTraitCollection: nil)?.imageWithRenderingMode(.AlwaysTemplate)
-    
-    func configure(cell: UITableViewCell) {
-        guard let fingerprintCell = cell as? ZomFingerprintCell else {
-            return
-        }
-        fingerprintCell.shareButton.setImage(self.shareImage, forState: .Normal)
-        fingerprintCell.qrButton.setImage(UIImage(named: "zom_qrcode_placeholder", inBundle: OTRAssets.resourcesBundle(), compatibleWithTraitCollection: nil), forState: .Normal)
-        fingerprintCell.fingerprintLabel.text = fingerprint
-        fingerprintCell.qrAction = {cell in
-            if let action = self.qrAction {
-                action(info:self)
-            }
-        }
-        fingerprintCell.shareAction = {cell in
-            if let action = self.shareAction {
-                action(info:self)
-            }
-        }
-    }
-    
-    func cellIdentifier() -> ZomProfileViewCellIdentifier {
-        return .FingerprintCell
-    }
-    
-    func cellHeight() -> CGFloat? {
-        return 90
-    }
-}
-
-struct ButtonCellInfo: ZomProfileViewCellInfoProtocol {
-    
-    enum ButtonCellType {
-        case Verify
-        case Refresh
-        case StartChat
-        
-        func text() -> String {
-            switch self {
-            case .Verify : return NSLocalizedString("Verify Contact", comment: "Button label to verify contact security")
-            case .Refresh: return NSLocalizedString("Refresh Session", comment: "Button label to refresh an OTR session")
-            case .StartChat: return NSLocalizedString("Start Chat", comment: "Button label to start a chat")
-            }
-        }
-    }
-    
-    let type:ButtonCellType
-    
-    func configure(cell:UITableViewCell) {
-        cell.textLabel?.text = self.type.text()
-        cell.textLabel?.textColor = UIButton(type: .System).titleColorForState(.Normal)
-    }
-    func cellIdentifier() -> ZomProfileViewCellIdentifier {
-        return .ButtonCell
-    }
-    func cellHeight() -> CGFloat? {
-        return nil
-    }
-}
-
-struct PasswordCellInfo: ZomProfileViewCellInfoProtocol {
-    let password:String
-    
-    func configure(cell: UITableViewCell) {
-        guard let passwordCell = cell as? ZomPasswordCell else {
-            return
-        }
-        passwordCell.passwordTextField.text = self.password
-        
-        passwordCell.changeButton.titleLabel?.font = UIFont(name: "FontAwesome", size: 30)
-        passwordCell.changeButton.setTitle(NSString.fa_stringForFontAwesomeIcon(.FAEdit), forState: UIControlState.Normal)
-        passwordCell.revealButton.titleLabel?.font = UIFont(name: "FontAwesome", size: 30)
-        passwordCell.revealButton.setTitle(NSString.fa_stringForFontAwesomeIcon(.FAEye), forState: UIControlState.Normal)
-        passwordCell.selectionStyle = .None
-    }
-    
-    func cellIdentifier() -> ZomProfileViewCellIdentifier {
-        return .PasswordCell
-    }
-    
-    func cellHeight() -> CGFloat? {
-        return nil
-    }
-}
-
 class ZomProfileTableViewSource:NSObject, UITableViewDataSource, UITableViewDelegate {
     
-    var info:ZomProfileViewControllerInfo
+    let info:ZomProfileViewControllerInfo
     var controller:ZomProfileViewController
     var relaodData:(() -> Void)?
     
@@ -392,9 +231,13 @@ class ZomProfileTableViewSource:NSObject, UITableViewDataSource, UITableViewDele
     }
 }
 
-class ZomProfileViewController : UIViewController, OTRAttachmentPickerDelegate {
+public class ZomProfileViewController : UIViewController {
     
     private var avatarPicker:OTRAttachmentPicker?
+    private var readOnlyDatabaseConnection:YapDatabaseConnection {
+        return OTRDatabaseManager.sharedInstance().readOnlyDatabaseConnection
+    }
+    private var viewHandler:OTRYapViewHandler?
     
     let tableView = UITableView(frame: CGRectZero, style: .Grouped)
     private var tableViewSource:ZomProfileTableViewSource?
@@ -419,6 +262,11 @@ class ZomProfileViewController : UIViewController, OTRAttachmentPickerDelegate {
             self.tableView.delegate = self.tableViewSource
             self.tableView.dataSource = self.tableViewSource
             self.tableView.reloadData()
+            
+            if let yapKey = self.info?.user.yapKey(), yapCollection = self.info?.user.yapCollection() {
+                //TODO: Remove all previous key collection pairs
+                self.viewHandler?.keyCollectionObserver.observe(yapKey, collection: yapCollection)
+            }
         }
     }
     var qrAction:((FingerprintCellInfo) -> Void)?
@@ -464,11 +312,11 @@ class ZomProfileViewController : UIViewController, OTRAttachmentPickerDelegate {
         }
     }
     
-    required init?(coder aDecoder: NSCoder) {
+    required public init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    override func viewDidLoad() {
+    override public func viewDidLoad() {
         super.viewDidLoad()
         
         self.tableView.translatesAutoresizingMaskIntoConstraints = false
@@ -486,6 +334,9 @@ class ZomProfileViewController : UIViewController, OTRAttachmentPickerDelegate {
         
         self.view.addSubview(self.tableView)
         self.tableView.autoPinEdgesToSuperviewEdges()
+        
+        self.viewHandler = OTRYapViewHandler(databaseConnection: self.readOnlyDatabaseConnection)
+        self.viewHandler?.delegate = self
     }
     
     @IBAction func didPressChangePasswordButton(sender: UIButton) {
@@ -554,31 +405,73 @@ class ZomProfileViewController : UIViewController, OTRAttachmentPickerDelegate {
     @IBAction func didTapAvatarImageWithSender(sender: UIButton) {
         if let user = self.info?.user {
             switch user {
-            case let .Account(_):
+            case .Account(_):
                 // Keep strong reference
                 avatarPicker = OTRAttachmentPicker(parentViewController: self.tabBarController?.navigationController, delegate: self)
                 avatarPicker!.showAlertControllerWithCompletion(nil)
                 break
-                default:
+            default:
                 break
             }
         }
     }
     
+    
+}
+
+extension ZomProfileViewController: OTRAttachmentPickerDelegate {
     public func attachmentPicker(attachmentPicker: OTRAttachmentPicker!, gotVideoURL videoURL: NSURL!) {
-        print("Got a video!")
+        
     }
     
     public func attachmentPicker(attachmentPicker: OTRAttachmentPicker!, gotPhoto photo: UIImage!, withInfo info: [NSObject : AnyObject]!) {
         if let user = self.info?.user {
             switch user {
             case let .Account(account):
-                account.avatarData = UIImagePNGRepresentation(photo)
+                if let xmppManager = OTRProtocolManager.sharedInstance().protocolForAccount(account) as? OTRXMPPManager {
+                    xmppManager.setAvatar(photo, completion: { [weak self] (success) in
+                        self?.tableViewSource?.relaodData?()
+                        })
+                }
                 break
             default:
                 break
             }
         }
-        print("Got a photo!")
+    }
+    
+    public func attachmentPicker(attachmentPicker: OTRAttachmentPicker!, preferredMediaTypesForSource source: UIImagePickerControllerSourceType) -> [String]! {
+        return [kUTTypeImage as String]
+    }
+}
+
+extension ZomProfileViewController: OTRYapViewHandlerDelegateProtocol {
+    public func didReceiveChanges(handler: OTRYapViewHandler, key: String, collection: String) {
+        
+        guard let info = self.info else {
+            return
+        }
+        
+        //The User object has changed. New info on the buddy and account
+        var newObject:OTRYapDatabaseObject?
+        self.readOnlyDatabaseConnection.readWithBlock { (transaction) in
+            newObject = transaction.objectForKey(key, inCollection: collection) as? OTRYapDatabaseObject
+        }
+        var user:ZomProfileViewControllerInfo.User?
+        switch newObject {
+        case let account as OTRAccount:
+            user = ZomProfileViewControllerInfo.User.Account(account)
+            break
+        case let buddy as OTRBuddy:
+            user = ZomProfileViewControllerInfo.User.Buddy(buddy)
+            break
+        default: break
+        }
+        
+        guard let newUser = user else {
+            return
+        }
+        
+        self.info = ZomProfileViewControllerInfo(tableSections: info.tableSections, user: newUser, otrKit: info.otrKit, otrKitInfo: info.otrKitInfo, hasSession: info.hasSession)
     }
 }
